# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * GOAT-PAY API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardSummary {
  totalRevenue: number;
  totalSales: number;
  avgTicket: number;
  conversionRate: number;
  refunds: number;
  revenueGrowth: number;
  salesGrowth: number;
  ticketGrowth: number;
  conversionGrowth: number;
  refundGrowth: number;
}

export interface RevenuePoint {
  date: string;
  revenue: number;
  sales: number;
}

export interface TrafficSource {
  source: string;
  percentage: number;
  count: number;
}

export interface TopProduct {
  id: number;
  name: string;
  revenue: number;
  sales: number;
  rank: number;
}

export interface WalletSummary {
  available: number;
  pending: number;
  total: number;
}

export interface Product {
  id: number;
  name: string;
  description?: string;
  type: string;
  price: number;
  status: string;
  sales: number;
  revenue: number;
  imageUrl?: string;
  createdAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  type: string;
  price: number;
  status?: string;
  imageUrl?: string;
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  type?: string;
  price?: number;
  status?: string;
  imageUrl?: string;
}

export interface Sale {
  id: number;
  productId: number;
  productName: string;
  customerName: string;
  customerEmail: string;
  amount: number;
  status: string;
  paymentMethod: string;
  country?: string;
  affiliateId?: number;
  createdAt: string;
}

export interface Transaction {
  id: number;
  type: string;
  description: string;
  amount: number;
  balance?: number;
  saleId?: number;
  createdAt: string;
}

export interface Withdrawal {
  id: number;
  amount: number;
  status: string;
  pixKey: string;
  notes?: string;
  createdAt: string;
  updatedAt: string;
}

export interface WithdrawalInput {
  amount: number;
  /** @minLength 1 */
  pixKey: string;
  notes?: string;
}

export interface Affiliate {
  id: number;
  name: string;
  email: string;
  commissionRate: number;
  totalSales: number;
  totalCommission: number;
  pendingCommission?: number;
  status: string;
  createdAt: string;
}

export interface AffiliateStats {
  totalAffiliates: number;
  activeAffiliates: number;
  totalCommissionPaid: number;
  totalSalesGenerated: number;
  topAffiliateName: string;
  topAffiliateRevenue: number;
}

export type ListProductsParams = {
type?: string;
status?: string;
};

export type ListSalesParams = {
status?: string;
productId?: number;
};