# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GOAT-PAY API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: dashboard
  - name: products
  - name: sales
  - name: wallet
  - name: withdrawals
  - name: affiliates

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      responses:
        "200":
          description: Dashboard KPIs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/revenue-chart:
    get:
      operationId: getRevenueChart
      tags: [dashboard]
      responses:
        "200":
          description: Revenue chart
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenuePoint"

  /dashboard/traffic-sources:
    get:
      operationId: getTrafficSources
      tags: [dashboard]
      responses:
        "200":
          description: Traffic sources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrafficSource"

  /dashboard/top-products:
    get:
      operationId: getTopProducts
      tags: [dashboard]
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"

  /dashboard/wallet-summary:
    get:
      operationId: getWalletSummary
      tags: [dashboard]
      responses:
        "200":
          description: Wallet summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletSummary"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /sales:
    get:
      operationId: listSales
      tags: [sales]
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: productId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"

  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
        "404":
          description: Not found

  /sales/{id}/refund:
    post:
      operationId: refundSale
      tags: [sales]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Refunded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"

  /wallet/transactions:
    get:
      operationId: listTransactions
      tags: [wallet]
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"

  /withdrawals:
    get:
      operationId: listWithdrawals
      tags: [withdrawals]
      responses:
        "200":
          description: Withdrawals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Withdrawal"
    post:
      operationId: requestWithdrawal
      tags: [withdrawals]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"

  /withdrawals/{id}/approve:
    post:
      operationId: approveWithdrawal
      tags: [withdrawals]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"

  /affiliates:
    get:
      operationId: listAffiliates
      tags: [affiliates]
      responses:
        "200":
          description: Affiliates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Affiliate"

  /affiliates/stats:
    get:
      operationId: getAffiliateStats
      tags: [affiliates]
      responses:
        "200":
          description: Affiliate stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AffiliateStats"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    DashboardSummary:
      type: object
      required: [totalRevenue, totalSales, avgTicket, conversionRate, refunds, revenueGrowth, salesGrowth, ticketGrowth, conversionGrowth, refundGrowth]
      properties:
        totalRevenue:
          type: number
        totalSales:
          type: integer
        avgTicket:
          type: number
        conversionRate:
          type: number
        refunds:
          type: number
        revenueGrowth:
          type: number
        salesGrowth:
          type: number
        ticketGrowth:
          type: number
        conversionGrowth:
          type: number
        refundGrowth:
          type: number

    RevenuePoint:
      type: object
      required: [date, revenue, sales]
      properties:
        date:
          type: string
        revenue:
          type: number
        sales:
          type: integer

    TrafficSource:
      type: object
      required: [source, percentage, count]
      properties:
        source:
          type: string
        percentage:
          type: number
        count:
          type: integer

    TopProduct:
      type: object
      required: [id, name, revenue, sales, rank]
      properties:
        id:
          type: integer
        name:
          type: string
        revenue:
          type: number
        sales:
          type: integer
        rank:
          type: integer

    WalletSummary:
      type: object
      required: [available, pending, total]
      properties:
        available:
          type: number
        pending:
          type: number
        total:
          type: number

    Product:
      type: object
      required: [id, name, type, price, status, sales, revenue, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        type:
          type: string
        price:
          type: number
        status:
          type: string
        sales:
          type: integer
        revenue:
          type: number
        imageUrl:
          type: string
        createdAt:
          type: string

    ProductInput:
      type: object
      required: [name, type, price]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        type:
          type: string
        price:
          type: number
        status:
          type: string
        imageUrl:
          type: string

    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
        price:
          type: number
        status:
          type: string
        imageUrl:
          type: string

    Sale:
      type: object
      required: [id, productId, productName, customerName, customerEmail, amount, status, paymentMethod, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        customerName:
          type: string
        customerEmail:
          type: string
        amount:
          type: number
        status:
          type: string
        paymentMethod:
          type: string
        country:
          type: string
        affiliateId:
          type: integer
        createdAt:
          type: string

    Transaction:
      type: object
      required: [id, type, description, amount, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
        description:
          type: string
        amount:
          type: number
        balance:
          type: number
        saleId:
          type: integer
        createdAt:
          type: string

    Withdrawal:
      type: object
      required: [id, amount, status, pixKey, createdAt, updatedAt]
      properties:
        id:
          type: integer
        amount:
          type: number
        status:
          type: string
        pixKey:
          type: string
        notes:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string

    WithdrawalInput:
      type: object
      required: [amount, pixKey]
      properties:
        amount:
          type: number
        pixKey:
          type: string
          minLength: 1
        notes:
          type: string

    Affiliate:
      type: object
      required: [id, name, email, commissionRate, totalSales, totalCommission, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        commissionRate:
          type: number
        totalSales:
          type: integer
        totalCommission:
          type: number
        pendingCommission:
          type: number
        status:
          type: string
        createdAt:
          type: string

    AffiliateStats:
      type: object
      required: [totalAffiliates, activeAffiliates, totalCommissionPaid, totalSalesGenerated, topAffiliateName, topAffiliateRevenue]
      properties:
        totalAffiliates:
          type: integer
        activeAffiliates:
          type: integer
        totalCommissionPaid:
          type: number
        totalSalesGenerated:
          type: integer
        topAffiliateName:
          type: string
        topAffiliateRevenue:
          type: number