# lib/api-zod/src/generated/types/affiliate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * GOAT-PAY API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Affiliate {
  id: number;
  name: string;
  email: string;
  commissionRate: number;
  totalSales: number;
  totalCommission: number;
  pendingCommission?: number;
  status: string;
  createdAt: string;
}