# lib/api-zod/src/generated/types/affiliateStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * GOAT-PAY API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AffiliateStats {
  totalAffiliates: number;
  activeAffiliates: number;
  totalCommissionPaid: number;
  totalSalesGenerated: number;
  topAffiliateName: string;
  topAffiliateRevenue: number;
}