# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * GOAT-PAY API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalRevenue: number;
  totalSales: number;
  avgTicket: number;
  conversionRate: number;
  refunds: number;
  revenueGrowth: number;
  salesGrowth: number;
  ticketGrowth: number;
  conversionGrowth: number;
  refundGrowth: number;
}