# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * GOAT-PAY API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  description?: string;
  type: string;
  price: number;
  status: string;
  sales: number;
  revenue: number;
  imageUrl?: string;
  createdAt: string;
}