# lib/api-zod/src/generated/types/revenuePoint.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * GOAT-PAY API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RevenuePoint {
  date: string;
  revenue: number;
  sales: number;
}